+++ src/answer_eval.py
@@ -1,8 +1,7 @@
 from models.answer import AnswerInput, AnswerResults
-from supabase.client import Client
 from transformers import logging
 from pipelines.answer import AnswerPipeline
-from src.database import get_client
+from src.database import Strapi, get_strapi
 
 logging.set_verbosity_error()
 
@@ -10,26 +9,12 @@ answer_pipe = AnswerPipeline()
 
 
 class Answer:
-    def __init__(self, answer_input: AnswerInput, db: Client):
-        # TODO: Change to use section slug
-        # This process should be the same for all textbooks.
-        if answer_input.textbook_name.name == "THINK_PYTHON":
-            section_index = f"{answer_input.chapter_index:02}"
-        elif answer_input.textbook_name.name in ["MACRO_ECON", "MATHIA"]:
-            section_index = (
-                f"{answer_input.chapter_index:02}-{answer_input.section_index:02}"
-            )
-        else:
-            raise ValueError("Textbook not supported.")
+    def __init__(self, answer_input: AnswerInput, db: Strapi):
+        response = db.fetch(
+            f"/api/pages?populate[Content][filters][slug][$eq]={AnswerInput.chunk_slug}"
+        )
 
-        self.data = (
-            db.table("subsections")
-            .select("clean_text", "question", "answer")
-            .eq("section_id", section_index)
-            .eq("subsection", answer_input.subsection_index)
-            .execute()
-            .data
-        )[0]
+        self.content = response["data"][0]["attributes"]["Content"]
 
         self.answer = answer_input.answer
         self.results = {}
@@ -39,7 +24,7 @@ class Answer:
         This currently returns passing score ONLY if both BLEURT and MPnet agree that it is passing
         """
 
-        correct_answer = self.data["answer"]
+        correct_answer = self.content["ConstructedResponse"]
         res = answer_pipe(self.answer, correct_answer)
 
         self.results["score"] = res
@@ -50,7 +35,7 @@ class Answer:
 
 
 async def answer_score(answer_input: AnswerInput) -> AnswerResults:
-    db = get_client(answer_input.textbook_name)
+    db = get_strapi()
 
     answer = Answer(answer_input, db)
     answer.score_answer()
