+++ src/main.py
@@ -9,8 +9,9 @@ from models.embedding import ChunkInput, ChunkEmbedding
 from models.chat import ChatInput, ChatResult
 from models.transcript import TranscriptInput, TranscriptResults
 
-from src.summary_eval import summary_score
-from src.summary_eval_strapi import summary_score_strapi
+from src.summary_eval_supabase import summary_score_supabase
+from src.summary_eval import summary_score_strapi
+from src.answer_eval_supabase import answer_score_supabase
 from src.answer_eval import answer_score
 from src.retrieve import generate_embedding
 from src.get_transcript import generate_transcript
@@ -53,14 +54,17 @@ def gpu_available():
 @app.post("/score/summary")
 async def score_summary(input_body: SummaryInput) -> SummaryResults:
     if input_body.textbook_name:
-        return await summary_score(input_body)
-    else:
         return await summary_score_strapi(input_body)
+    else:
+        return await summary_score(input_body)
 
 
 @app.post("/score/answer")
 async def score_answer(input_body: AnswerInput) -> AnswerResults:
-    return await answer_score(input_body)
+    if input_body.textbook_name:
+        return await answer_score_supabase(input_body)
+    else:
+        return await answer_score(input_body)
 
 
 @app.post("/generate/embedding")
